requirements.txt
absl-py==0.14.0
APScheduler==3.6.3
astroid==2.7.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.1.3
flake8==3.9.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
libretranslatepy==2.1.1
lxml==4.9.0
Markdown==3.3.4
mccabe==0.6.1
numpy==1.21.2
oauthlib==3.1.1
platformdirs==2.3.0
protobuf==3.18.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.10.2
Pyrogram==2.0.26
PySocks==1.7.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
toml==0.10.2
tornado==6.1
translate==3.6.1
tzlocal==3.0
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
